increase pseudoKeyStretching security
move from hex to base64 encoding. Provides larger charset.

WARNING: breaks all data stored using hex fixed keys.
As there are no systems in production yet, it should be mostly fine.

This change could be issued as another property on the Constructor,
but this is opnionated code, and better standards should be enforced
as it increases security. And reduce a chance of misconfiguration.